\name{modsemTools-package}
\alias{modsemTools-package}
\alias{modsemTools}
\docType{package}
\title{
\packageTitle{modsemTools}
}
\description{
\packageDescription{modsemTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{modsemTools}
\packageIndices{modsemTools}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{modsemTools}

Maintainer: \packageMaintainer{modsemTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
